Move files into authentication package
