+++ src/main/kotlin/com/vandoc/model/request/authentication/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.vandoc.model.request
+package com.vandoc.model.request.authentication
 
 data class LoginRequest(
     val email: String?,
